cryptohome: Change PW lockout action to LeLockedOut
Currently, LECredentialManager returns TpmLockout when the leaf is
locked out. Now that we have a designated error code for LE lockout, we
should use that instead. In summary, when an incorrect PIN attempt
results in lockout, both kIncorrectAuth and kLeLockedOut should be
returned. When a PIN attempt failed before matching because the leaf is
already locked out, kLeLockedOut should be returned.

BUG=b:269985920
TEST=cryptohome unittests

Change-Id: Ie7ca10b994e20e9827332cfe247cb8817e58bb9f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4290541
Reviewed-by: John L Chen <zuan@chromium.org>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>