cryptohome: Allow sessions to be added to the session manager
This is useful in testing where there are cases where we'd like to be
able to more directly control session creation. The manager also uses
this new Add directly in Create, so that behavior is consistent. For
simplicity I did turn the token collision failure into a CHECK-fail, as
these tokens are supposed to be secure, this really should be a "never
happens".

BUG=b:263259501
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I16abd2995b86789ffe6804edbaafe40bd08294f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4292296
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>