shill: Consolidate IPAddress usage (part 2)
This series of patches mainly replace the usage of IPAddress ctor to
IPAddress::CreateFrom*() factory methods, so that 1) it's more obvious
that the construction of IPAddress can fail, and 2) the IPAddress object
created by the factory methods will always be valid, so IsValid() is no
longer needs to check.

This series of patches should not affect the current code logic.

BUG=b:269983153
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: I838dacd4c20f6c56ddd0407569938925e53449a2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4290532
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>