installer: Fix possible mount fail in ESPPostInstall
There is a race condition between:
1) `udevd`, which is indirectly and, more importantly, asynchronously
   invoked by the previous call to `UpdatePartitionTable`; and
2) the `mount` call in `ESPPostInstall`.
`UpdatePartitionTable` modifies the drive (let us use `/dev/sda` as
the drive for example's sake but this applies to other drives too,
including NVMe drives) containing the EFI system partition (ESP) that
ESPPostInstall will call `mount` on. The drive is being watched due to a
`udev` rule in `/lib/udev/rules.d/60-block.rules` that applies
`OPTIONS+="watch"`. This makes `udevd` call `synthesize_change` (found
in systemd's `udevd.c`) when the drive is modified in
`UpdatePartitionTable`. In `synthesize_change`, the partitions of the
drive (e.g. `/dev/sdaX` for X from 1 to 12) are re-read via a call to
`ioctl`. This re-adds the device nodes, for the partitions, but first
removes them. This `synthesize_change` is called in `udevd`; it is not
called by the code here and so we do not wait on `udevd` during
the call to `UpdatePartitionTable`. Meanwhile, we try to `mount` the
ESP (located at `/dev/sda12` in our example) in the call to
`ESPPostInstall` below, but if the timing is wrong, the `mount`
will fail to find the ESP (located at `/dev/sda12` in our example)
because it might be removed while in the process of being re-added by
the `ioctl` call, to re-read partitions, made by`udevd`. Running
`udevadm settle` resolves this issue by making sure that re-reading
the partitions is done before we call `mount` on the ESP in
`ESPPostInstall`.

BUG=b:263502368, b:264571236, b:265487884, b:268521748
TEST=Use a USB drive to install Flex on the affected machines and verify
     that there is no error message saying "Failed to mount /dev/sda12
     to /tmp/boot_mnt: No such file or directory" during installation.

Change-Id: I5ec83ade60c100c76532ee22c8fe11d917392178
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4296093
Commit-Queue: Adrian Chan <adchan@google.com>
Tested-by: Adrian Chan <adchan@google.com>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>