cryptohome: func GetAuthBlockTypeForCreation signature change.
Prior to the introduction of auth factor, we needed to
create a list of auth blocks based on user inputs. The original
GetAuthBlockTypeForCreation was operating on a list of
booleans. Now it is more and more difficult to use when we
introduce new auth factors. Therefore, it is good to
convert it to use AuthFactorType with a switch statement.

BUG=b:270076334
TEST=CQ

Change-Id: I356ac8f52b00ddf3b75a61eade6f3d2fbadf23ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4297767
Commit-Queue: Zi Lin <lziest@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Zi Lin <lziest@google.com>