u2fd: system_api: Fix IsUvpaa method behavior
As WebAuthn using password is supported, IsUvpaa should be true as long
as the user is logged in and WebAuthnHandler is initialized.

Use a new proto field so that Chrome will be able to distinguish and
rely on the new behavior.

BUG=b:270509867
TEST=u2fd unittests

Change-Id: I8d3b5b562f7f41119f3c7b1b5d39ca6a01c143af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4296859
Reviewed-by: Yi Chou <yich@google.com>
Auto-Submit: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>