cryptohome: Add rate-limiter reset secret in USS
The existing reset_secret field in the USS payload is indexed by the
AuthFactor label. For rate-limiter, it doesn't have a corresponding
label because it's for the whole AuthFactor type. It will be too much of
a workaround to use some placeholder string as the label name. Instead,
introduce a new field to put reset secrets for rate-limiters, indexed by
AuthFactor type.

BUG=b:270076334
TEST=cryptohome unittests

Change-Id: I865818a2815dc605cdcb5e9e9d5bee49cb56b703
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4302550
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>