shill: cellular: update active bearers upon a successful connection
The list of active bearers should be updated any time a connection
attempt has successfully finished, and not only when the modem
transitions into "connected" state, given that this transition will
happen exclusively for the first connected bearer and not for the next
ones.

At the same time, avoid the early connection setup happening upon the
state update and instead explicitly wait for the attempt to be
reported as successful.

BUG=b:271470932
TEST=CQ passes

Change-Id: I1028cca6f69e2a2ee82c9a2a199a6f20d010b4db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4304301
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>