login Only configure chrome cgroups once (redux)
Today ui-pre-start is run at each login. This means we're attempt to
create and reconfigure chrome cgroups every time. This prevents
featured from configuring them as part of an experiment. This change
only writes to the uclamp files once when the cgroup is created.

NOTE: this is the 2nd time landing this same CL, but this time it
include a bugfix (hence the "redux" in the subject). In the original
CL <https://crrev.com/c/4263027> ("login: Only configure chrome
cgroups once") we accidentally broke "restart ui" for boards using
schedutil because they were relying on the value of ${v} that was only
populated the first time the script was run.

Let's fix the problem by moving the schedutil setting together with
uclamp setting. We'll also add a bug reference to the newly-filed bug
about when we can remove the old schedutil stuff.

BUG=b:248028313, b:271361813
TEST=Put on DUT and see things configured correctly

Change-Id: I04b58f7cac87581c43cc2c327971a8018a55b619
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4302695
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>