cryptohome: Get rid of AuthBlockType::kMaxValue
This simplifies a bunch of code in cases where we need to have an
unnecessary error path to handle kMaxValue. In cases where we do need to
have an failure path, we use optional<AuthBlockType> to indicate this.

For metrics it's still useful to have a well defined max value, because
you need to supply a min and max with UMA metrics. However, we don't
need an enum entry for this, we can just define a separate max value out
of line instead.

BUG=b:237115296
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Idfd073a61f02d6688489e1270858f2e798ddeb7b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4302698
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>