arc-setup: Remount shared data directory as NOEXEC
In the init namespace, we have a scenario where
/run/arc/shared-mounts/data is both RW and EXEC.
This is a vulnerability. Ultimately, ARC container
needs the EXEC capability, but we want the above
directory to be NO_EXEC only.

The above mount point is first created, passed to ARC and
then unmounted. Long time window increases the vulnerability time
leading it to be detected by secanomalyd.

This CL reduces that window by remounting it as NO_EXEC in the
init namespace soon after the shared mount is passed to ARC.
In the ARC namespace, the shared mount still has the
required EXEC capability.

BUG=b:219574442
TEST=arc.Boot
TEST=Ensure that /data directory inside the ARC container
has EXEC capability.

Change-Id: I98399c468c6d9d82d52698bf9e8bfd423f5467e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4305719
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Yury Khmel <khmel@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Yury Khmel <khmel@google.com>