patchpanel: Match output interface in vpn_accept rule
The ACCEPT rule in vpn_accept chain is used to accept the packets should
be sent via VPN, so that they will not be dropped by the REJECT rule in
vpn_reject chain in the always-on vpn strict mode.

When the VPN is disconnected, the routes and ip rules will be removed at
first by shill, and iptables rules will be removed by patchpanel after
that. This causes an issue that, during this time window, in strict
mode, the routing system will decide to route a packet to the physical
interface (since routes for VPN have been removed), and iptables accept
this packet (since iptables rules in vpn_accpet chain is still there),
and thus the packet will go out from the physical interface, which is
not excepted.

This patch fixes this issue by letting the ACCEPT rule in vpn_accept
chain also match the output interface, so that it can be guaranteed that
a packet routed to the physical interface will not be accepted by this
rule.

BUG=b:269431032
TEST=FEATURES="test" emerge-$BOARD patchpanel
TEST=`tast run $DUT network.AlwaysOnVPNRouting.*` passed (note that this
  test is just flaky before this patch, so this verification may not be
  enough).

Change-Id: I576ddb3f0a4aa70369db424783edac4883f5166b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4305899
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>