shill: Refactor ScanAndConnectToBestServices.
ScanAndConnectToBestServices only scans on the WiFi interface. If the
best service is on another technology, ScanAndConnectToBestServices
shouldn't block connections to services on other technologies on the
WiFi scan.  Additionally, it's clearer to have the WiFi device only
responsible for ensuring that the best WiFi service; there are likely
latent bugs around edge cases where ConnectToBestServices doesn't get
called but should be as the result of a bug in the WiFi Device.

In order to accomplish that goal, the following changes are made:
* Split ConnectToBestServices into ConnectToBestWiFiService and
  ConnectToBestNonWiFiServices.
* Call EnsureScanAndConnectToBestService when a WiFi device is found,
  and call ConnectToBestNonWiFiServices whether the WiFi device is found
  or not.
* Test ScanAndConnectToBestServices.

BUG=b:270969976

TEST=FEATURES="test" USE="-cros-debug" emerge-skyrim shill

Change-Id: I5a485dc89cfbe7012ad21cc55ad0758f10eab020
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4307011
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: David Ruth <druth@chromium.org>
Tested-by: David Ruth <druth@chromium.org>