printscanmgr: switch to brillo::ProcessImpl from ProcessWithOutput
The main functionality that ProcessWithOutput gives is the ability to
sandbox the process. This makes sense in debugd, which is running as
root. For the main printscanmgr daemon, the entire daemon process is
already sandboxed, so we don't need additional sandboxing on the
commands that we run.

BUG=b:257070388
TEST=FEATUES=test emerge-nami printscanmgr

Change-Id: Icc3dacd3e2e4e96d69fdd92bb350c3ba6d0af354
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4307562
Commit-Queue: Paul Moy <pmoy@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Paul Moy <pmoy@chromium.org>