ml_core: handle empty cache startup in camera effects
When the UI restarts after a chrome flag change, the opencl_cacher tool
will not be run (it happens earlier in the boot cycle). This means the
cache directory will be empty, and the user will experience a ~10-15s
delay for effects to turn on.

In this case, and when the effects marker file indicating a crash is NOT
present, then point the pipeline at the prebuilt cache directory to
handle this edge case.

With this new behaviour, the opencl_cacher tool does not need to copy
the prebuilt cache files into the cache directory, so removing this
logic.

BUG=b:257160623
TEST=build, deploy to brya. Test behaviour after flag set and reboot.

Change-Id: I52a9b9c652be8e474545e5e9c7cd7267427673d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4313579
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>
Tested-by: Jim Pollock <jmpollock@chromium.org>