secagentd: Add a Batch message sender
This is a helper for coalescing events that are destined to a single
Destination. It takes care of periodically flushing enqueued events as
well as flushing immediately if the batch size approaches an ERP
specific threshold. In addition, it also provides a functionality to
lookup and mutate pending events with a caller provided function.

Heavily templated because the idea is that this class can be reused
across destinations as long as the protos follow the same general
nesting pattern.

Will be connected to the daemon in a followup cl. Only unit test
coverage for now.

BUG=b:271918420
TEST=All secagentd unit tests pass

Change-Id: I42b400df40788c137055bc4fb735b3735e95771d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4316034
Tested-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Ryan Borzello <rborzello@google.com>
Reviewed-by: Ben Scarlato <akhna@google.com>