cryptohome: Remove backup creation code
The CL removes uage of enable_create_backup_vk_with_uss_. A vault keyset
is only added/updated/removed if there is at least one stored factor that
is backed by vault keyset.

Further, if during AuthSession creation we find that we are in midst of
USS AuthFactors, we remove the backups as those are not required
anymore.

BUG=b:271976462
TEST=CQ passes

Change-Id: I135aea6a7bf7564a3c4f78a9973a2454c4ae007e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4316382
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>