cryptohome: Make IsAuthBlockSupported generic
This is an attempt at pulling one of the auth block utility "select a
type and call it" functions out into a more generic implementation that
doesn't require manually maintained switches or if-else chains. This
does come at the cost of some template magic, but said magic has the
benefit of only needing to be written once per function, not updated as
AuthBlock implementations are changed.

For this initial version I keep some special-case code to handle
fingerprint, which had additional checks beyond IsSupported. That can
also be handled with some additional work, but I wanted to keep the
first version a little simpler. Once this function is finished it's not
to hard to apply similar patterns to a lot of the other functions.

BUG=b:272098290
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ic7eae9b695ce3ced75e0b0601027379ea3eb9f90
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4317586
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>