hiberman: Don't abuse ActivatedLogicalVolume to enable LVs
setup_hibernate_lv() creates an ActivatedLogicalVolume and
calls alv.dont_deactivate() directly after. The whole purpose
of an ActivatedLogicalVolume is to deactivate the volume again
when the ALV is dropped, which means that an ALV isn't needed
here. Call activate_lv() instead of creating an ALV and calling
.dont_deactivate() on it.

Something similar happes with activate_thinpool_ro() and its
only caller. activate_thinpool_ro() returns an ALV and the
caller immediately calls .dont_deactivate() on it. Also change
that to an activate_lv() call.

BUG=b:266012455
TEST=hibernate works on brya

Change-Id: I16fa881f7dc0c2903a1c89a20cf8792bc9255477
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4323100
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>