secagentd: Plumb Batched event sender and feature broker
Adds a new feature flag that controls batching.
The batching feature flag is currently disabled and will be flipped in a
followup cl coupled with the required tast test changes.

Unit tests are updated for coverage but most of them still prefer the
deprecated path. The deprecated path is almost a superset of the new
path so there's no coverage gap. This will be cleaned up nonetheless in
the future cl that removes the deprecated path.

BUG=b:271918420
TEST=All secagentd unit and tast tests pass

Change-Id: Ibdda912dec4c1c5da5732cb34f61686dbb8e3265
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4316036
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Ryan Borzello <rborzello@google.com>
Tested-by: Aashay Shringarpure <aashay@google.com>