cryptohome: Make generic IsSupported more generic
This allows us to get rid of the special case for fingerprint in
IsAuthFactorSupported, by folding the "has biometrics service" check
into it as well.

THis does require making the generic thunk a little bit more clever:
instead of requiring a fixed set of IsSupported arguments in every auth
block, we can instead allow you to accept any subset of a predetermined
set of parameters. So the FingerprintAuthBlock::IsSupported can now have
an extra parameter that the others don't, and it all still works. The
implementation of this is actually simpler than you might think, with
the main disadvantage being that if you somehow mess up your types
you'll get a very very terrible compiler error and not something very
readable. :(

However, this technique is quite powerful as it can also be used to
refactor the other calls into generic calls as well, for example the
factory functions which by their very nature require different
parameters for each type.

BUG=b:272098290
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I9fea5713e141071275c23ef40be1d96fe01b2007
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4316030
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>