libbrillo: SafeFD ignore some ENOENT during recursive Rmdir
This does not return an error when ENOENT is encountered for contents
during a recursive Rmdir. We do not want ENOENT for contents being
conflated with ENOENT for the target directory. This should make this
less prone to failure when multiple processes are deleting overlapping
paths.

This also stops logging errors when Rmdir is called for a file since
that is an expected case in brillo::DeleteFile.

BUG=chromium:1398990
TEST=FEATURES=test emerge libbrillo

Change-Id: I3d8ceba935d3133f41c2837aff3caf4d4c636050
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4320545
Reviewed-by: Brian Norris <briannorris@chromium.org>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>