diagnostics: Refactor thunderbolt bus fetcher
Refactor the argument and the logic to find the interfaces.
Thunderbolt interfaces will be placed under the domain.
The original implementation try to use the symbolic link under
the same directory with the .../domainX path. It looks like
.../X-Y.
We can just find all the interface in the domainX directory. It
looks like .../domainX/X-Y.
Also remove the logic which check the interface name starts with X.

The original vendor / device name is fetched from one of the
interfaces under domainX directory. This can be changed to simply
reusing the data from the fetched interfaces.

BUG=b:272217834
TEST=FEATURES=test emerge-zork diagnostics

Change-Id: Ib421c57587255ed7bea9bb0c3f5b46e06f50bb57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4321508
Tested-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>