cryptohome: Get rid of sync createkeyblobswithauthblock
Trying to make the switched utility functions generic, this would be a
lot easier if we didn't have two versions of some of them. So lets get
rid of the old sync one, that's now only used in tests.

I converted the tests over to using the sync version, so that we don't
lose any practical coverage.

BUG=b:272098290
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ic0ef71a7f2da6e224373587a6fa40beb45f71ab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4321629
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>