secagentd: Opportunistically coalesce process terminate events
When a terminate event is to be enqueued, the process plugin tries to
find an exec event enqueued with the same uuid. If found (depending on
when the last flush was), it sets the terminate_timestamp_us field in
the same exec event and otherwise drops the terminate event.

Adds a new feature flag that controls this coalescing. This flag is
currently disabled and will be flipped in a followup cl coupled with the
required tast test changes.

BUG=b:272312732
TEST=All unit tests pass. Also tested manually on a vm.

Change-Id: I84edb430b2de45a21916fbea603410ce8b783e94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4321631
Commit-Queue: Aashay Shringarpure <aashay@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Ryan Borzello <rborzello@google.com>