hiberman: Change a bunch of types from i64 to u64
Change the type of some variables / parameters / return values that
are always positive from i64 to u64.

The timeout parameter of wait_for_snapshots_merge() is changed from
mut i64 to u32. This allows to convert it safely into a local u64
variable that is used to adjust the remaining timeout value, instead
of using the mut parameter for that.

BUG=b:266012455
TEST=hibernate works on brya

Change-Id: Ie996191f8326330fe1ccc815c747d88e0744295d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4323101
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>