cryptohome: Connect to biod's session failed signal
We previously didn't connect to the OnFinish method of the biod proxy.
It is more desirable for cryptohome to emit an error signal for the
PrepareAuthFactorProgress when the session failed, so Chrome can do
error handling. For that, we need to connect to the OnFinish method, to
know that something went wrong in the biod sessions.

BUG=b:249749541
TEST=cryptohome unittests

Change-Id: I652f4aabebd120a7103f5d90fe8cf3ed7072e604
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4323461
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>