cryptohome: Get rid of sync derivekeyblobswithauthblock
Trying to make the switched utility functions generic, this would be a
lot easier if we didn't have two versions of some of them. So lets get
rid of the old sync one, that's now only used in tests.

I converted the tests over to using the sync version, so that we don't
lose any practical coverage.

More usefully for what I'm trying to do, this also lets us get rid of
the sync factory function for auth blocks. So to genericize the auth
block construction, we only need to make it work with the async version
now.

BUG=b:272098290
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I1bdb1981edb0c9ae9f99362b3db454a1fd072623
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4324127
Reviewed-by: Behnood Momenzadeh <behnoodm@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>