debugd: Implement GetFeedbackLogsV3
GetFeedbackLogsV3 is the same as GetFeedbackLogsV2 except that it tries
to run sub task in parallel in order to reduce client wait time and to
reduce the chance of timing out.

The majority of logs are of type Log. The log data are generated either
by running a binary or by reading a log file(s).

- Fork one worker per log command. Let them all run to completion. Let
  the kernel figure out which ones can run when, depending on IO
  latency, etc etc.
- Each child process will write the log data to a unique temp file.
- When all children have exited or timing out, load log data from temp
  files whatever available.

There are a few other log types which have different interfaces. They
will be run in the main process so to let it do work while other forked
processes are running.

A feature flag #enable-get-debugd-log-in-parallel has been added in
a separate CL. When the flag is enabled, GetFeedbackLogsV3 will be used.

Design Doc: go/improve-getting-feedback-logs-from-debugd.

BUG=b:238782324
TEST=manually tested on DUTs.
 Volteer:  GetFeedbackLogsV2: ~8sec, GetFeedbackLogsV3: ~3sec.

Change-Id: Id22c55f8532cc9c30cfb4d737bfb72e8bc746ac3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4420207
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Xiangdong Kong <xiangdongkong@google.com>
Commit-Queue: Xiangdong Kong <xiangdongkong@google.com>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>