crosh: Deprecate builtin_corpssh command
Now that the corpssh feature is enabled by default and not reading this
flag, we can remove the implementation. However, for user experience,
keep this command there for a longer while, and modify the help message
to tell the users the deprecation of the command.

BUG=b:232334144
TEST=FEATURES=test emerge crosh
TEST=test on DUT that the command message is clear and correct

Change-Id: I0cdce8a1d6f2bc6958898c7a69c96d767d505450
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4328110
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>