cryptohome: data_migrator: Distinguish source/dest paths for metrics
Currently, when a migration by MigrationHelper fails due to a file
error, it only reports to the delegate the relative path from the
migration root to the file, and it is not always possible to know
whether the failure happened with a file in the migration source or the
destination. This patch therefore adds another argument to
ReportFailure() to report whether the file(s) that caused the failure is
in the migration source, the destination, or both.

BUG=b:272151802
TEST=FEATURES=test emerge-hatch cryptohome
TEST=Manually check with CL:4317704 that whether the failed path belongs
TEST=to source or the destination is correctly reported to the metrics.

Change-Id: Id0b4e4b936adf78f2a7ca7d9af8781403efc7ddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4328025
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>