cryptohome: Add CreateDirectoryAndGetError to Platform
The current implementation of Platform::CreateDirectory() does not set
errno correctly on failures, although many programs in cryptohome call
CreateDirectory() and reference errno when it fails.
This CL adds Platform::CreateDirectoryAndGetError(), a variant of
CreateDirectory() which also returns the error information, and uses it
from cryptohome::data_migrator to fix the error code reported to the
metrics.

BUG=b:272704660
TEST=FEATURES=test emerge-hatch cryptohome
TEST=Manually run ARCVM /data migration with a small space situation and
TEST=confirm that the error from Platform::CreateDirectory() is reported
TEST=to metrics as FILE_ERROR_NO_SPACE.

Change-Id: I33cc35bb00fae4a691191484e2b9589d227408e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4337116
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Daniil Lunev <dlunev@chromium.org>