shill: net: Define ctor/dtor of mock classes in cc files
The ctor and dtor of a mock class are non-trivial, and very expensive to
compile [1]. Defining them in the header file will let them be compiled
together with the test files, and thus make the compilation of a single
test file taking longer time. Since the compiler does not compile a
single translation unit parallelly, the whole build time for
shill_unittest is actually affected by the most expensive unit test file
(which is wifi_test.cc now), so reducing the compile time of a single
test file does matter here.

This series of patches move the definition of ctors and dtor of mock
classes in shill to separate cc files, so that they will be compiled
separately, and thus reduce the total build time of shill_unittests.

The following data shows the effect of these patches on the required
time for the whole build and for some large unit test files. Data are
collected by run `FEATURES="test" emerge-$BOARD shill` after changing
refptr_types.h. Values are in seconds.

     Total           66.3 -> 61.5
   wifi_test         54.3 -> 49.1
  manager_test       44.2 -> 43.7
 cellular_test       41.5 -> 38.9
device_info_test     39.3 -> 30.1
 ethernet_test       36.5 -> 31.7

[1] http://google.github.io/googletest/gmock_cook_book.html#making-the-compilation-faster

BUG=b:272384938
TEST=FEATURES="test" emerge-$BOARD shill-net shill

Change-Id: I696af3cdc5b13fe1a5c0d53f1969dd11f4bc73cf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4328511
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>