cryptohome: Make the GetAuthBlockWithType implementation generic
This adds a New function to each type that encapsulates the "make me an
instance of this type" operation. In all cases it basically just ends up
being a pretty simple validity check plus a call to the constructor, but
it moves the logic for each block into the block.

This required adding a bunch of new types into the generic handler, but
that was expected which is why I set it up the way I did. I ended up
having to make the parameters a mix of references and pointers; for
interfaces which should never be null I made them references, but for
ones which can be null if they're not available I made them into
pointers instead. Overall though the generic logic remains mostly the
same just with a lot of new options added for parameters to accept.

BUG=b:272098290
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I6bee4a75409b68972a7b09b665418414ed17fe16
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4336636
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>