featured: Stubs to save safe seed only if used.
If the safe seed sent by chrome does not match the seed featured used to
compute state, we *cannot* conclude that it is safe: consider the case
when featured first tries to use seed B, fails, falls back to safe seed
A, and then chrome starts up correctly. Chrome might mark seed B as safe
(since it successfully used it), but seed B is not safe for featured.

In general, featured should only mark a seed as safe if it used that
seed to boot.

BUG=b:273343808
TEST=CQ

Change-Id: Ie144ea1fc231ae610a02e6a442aa30913110a452
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4336207
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>