chromeos-config: Make CrosConfigJson deterministic
Some part of the code of CrosConfigJson relies on the field order of
`str(d)` for a dict `d`. To deterministically convert a dict to a
hashable, the suggested way is to use `json.dumps(d, sort_keys=True)`,
which also handles the case where the values of `d` are also
dictionaries.

Add DeviceConfigJson.GetPropertiesStr() method as a shortcut for string
conversion.

BUG=none
TEST=./run_tests.sh

Change-Id: I453fe378e49ef4ada0ab937d3fdb62f6440ac5cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4337659
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Auto-Submit: Yu-Ping Wu <yupingso@chromium.org>