libbrillo: Expose details of flag parsing failures.
In order to improve crash_reporter's resilience to unknown flags, we
want to allow it to continue to execute if it sees any unknown flags,
but *not* if it fails to parse known flags.

Expose information about the nature of Init failures so that when using
kReturn, callers may make decisions about what to do based on details of
which flag(s) were problems.

BUG=b:272345548
TEST=CQ

Change-Id: I134aec2d0e23a47b5ad76e6a8a6954c90f9d69ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4338405
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>