camera: Fix cros-camera-effects-sm-tests failures
At present the camera effects stream manipulator tests are failing and segfaulting. The segfault is caused by setting the config callback too early that the pipeline was not ready for loading the config when the callback is called. The tests are failing because the stream manipulator's stream was not configured correctly. Additionally, relight effect has changed since the test was originally written causing it to fail as well.

Loading the config after the pipeline is created to ensure the the pipeline is correctly configured. Correctly configure stream manipulator's streams to ensure effects are applied to the test stream. And update the relight reference image to match reality.

BUG=b:271941721
TEST=Running tests on DUT, CQ.

Cq-Depend: chromium:4337533
Change-Id: I8d50e756043ef523ef004a58dc3c919973382a25
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4337100
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Tested-by: Jingyuan Tu <zactu@google.com>
Commit-Queue: Jingyuan Tu <zactu@google.com>