shill: cellular: remove workaround in IPA network interface
shill no longer needs the hardcoded workaround that made the cellular
device object be bound to the rmnet_data0 interface even if rmnet_ipa0
is the net interface reported by ModemManager.

shill will now create the cellular device bound to rmnet_ipa0, as that
is the network interface reported by ModemManager in the list of
ports, and will then use for the actual network connection the
interface reported in the connected bearer.

BUG=b:249387022
TEST=Connectivity tests in trogdor and herobrine.

Change-Id: I0245ce4fb3d51c5cca958fc9a8e77e2cdbbe3434
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4340789
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>
Tested-by: Aleksander Morgado <aleksandermj@google.com>