cryptohome: Fix upstart script lint issues
For cryptohomed we should be able to just remove the mkdir command
entirely, because we already have a tmpfiles.d for cryptohome that sets
up /run/cryptohome (the recommended way to do it).

For mount-encrypted I instead disabled the lint warning. Ideally we
wouldn't do this, but in this case it's difficult (possibly impossible)
to exactly reproduce the semantics with tmpfiles.d (maybe an r-c-z
sequence?) and in this case the command seems to be reasonably safe to
execute:

- the -P will avoid following symlinks
- the src and dst directories are both root-owned

BUG=b:239059296
TEST=repohooks/pre-upload.py

Change-Id: I32adda47e50449fba7ae756e5091d08045f3053a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4343769
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Chris Palmer <palmer@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>